Move id provider internally, add unicode support
